migrate jupyter-serde into jupyter-protocol
