+++ crates/jupyter-protocol/src/media/mod.rs
